fix: :bug: Rename waitFor file
